requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==3.6.4.0
celery==5.2.7
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.14
Django==4.0.7
flower==1.2.0
humanize==4.9.0
kombu==5.3.5
packaging==23.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
pytz==2024.1
redis==4.3.4
sqlparse==0.4.4
tornado==6.4
typing_extensions==4.9.0
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
djangorestframework
django-cors-headers
psycopg2-binary